refactor: extract useGraphFilters composable (display computeds + language watch)
